+++ src/main/java/io/github/axolotlclient/modules/hud/gui/hud/vanilla/CrosshairHud.java
@@ -86,7 +86,7 @@ public class CrosshairHud extends AbstractHudEntry implements DynamicallyPositio
                     new int[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
                     new int[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
                     new int[]{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0},
-            });
+            }, true);
 
     public CrosshairHud() {
         super(15, 15);