+++ CHANGELOG.md
@@ -98,3 +98,4 @@
 
 - Add a `graphical` option type
 - Add custom crosshair texture option
+- add customization options to the mouse movement indicator of the KeystrokeHud